# api/generated/server.gen.go
// Package generated provides primitives to interact with the openapi HTTP API.
//
// Code generated by github.com/oapi-codegen/oapi-codegen/v2 version v2.6.0 DO NOT EDIT.
package generated

import (
	"fmt"
	"net/http"
	"time"

	"github.com/gin-gonic/gin"
	"github.com/oapi-codegen/runtime"
	openapi_types "github.com/oapi-codegen/runtime/types"
)

// Defines values for SakeCategory.
const (
	AWAMORI      SakeCategory = "AWAMORI"
	BEER         SakeCategory = "BEER"
	FRUITWINE    SakeCategory = "FRUIT_WINE"
	JAPANESESAKE SakeCategory = "JAPANESE_SAKE"
	LIQUEUR      SakeCategory = "LIQUEUR"
	NONALCOHOL   SakeCategory = "NON_ALCOHOL"
	OTHER        SakeCategory = "OTHER"
	SHOCHU       SakeCategory = "SHOCHU"
	WINE         SakeCategory = "WINE"
)

// Valid indicates whether the value is a known member of the SakeCategory enum.
func (e SakeCategory) Valid() bool {
	switch e {
	case AWAMORI:
		return true
	case BEER:
		return true
	case FRUITWINE:
		return true
	case JAPANESESAKE:
		return true
	case LIQUEUR:
		return true
	case NONALCOHOL:
		return true
	case OTHER:
		return true
	case SHOCHU:
		return true
	case WINE:
		return true
	default:
		return false
	}
}

// Defines values for ServiceUnavailableDatabasePostgres.
const (
	ServiceUnavailableDatabasePostgresError ServiceUnavailableDatabasePostgres = "error"
	ServiceUnavailableDatabasePostgresOk    ServiceUnavailableDatabasePostgres = "ok"
)

// Valid indicates whether the value is a known member of the ServiceUnavailableDatabasePostgres enum.
func (e ServiceUnavailableDatabasePostgres) Valid() bool {
	switch e {
	case ServiceUnavailableDatabasePostgresError:
		return true
	case ServiceUnavailableDatabasePostgresOk:
		return true
	default:
		return false
	}
}

// Defines values for ServiceUnavailableDatabaseValkey.
const (
	ServiceUnavailableDatabaseValkeyError ServiceUnavailableDatabaseValkey = "error"
	ServiceUnavailableDatabaseValkeyOk    ServiceUnavailableDatabaseValkey = "ok"
)

// Valid indicates whether the value is a known member of the ServiceUnavailableDatabaseValkey enum.
func (e ServiceUnavailableDatabaseValkey) Valid() bool {
	switch e {
	case ServiceUnavailableDatabaseValkeyError:
		return true
	case ServiceUnavailableDatabaseValkeyOk:
		return true
	default:
		return false
	}
}

// APIError defines model for APIError.
type APIError struct {
	// Code アプリケーション側で定義したエラーコード
	Code string `json:"code"`

	// Message ユーザー向けのエラー説明メッセージ
	Message string `json:"message"`
}

// CreateStockRequest 在庫登録時に送信するお酒情報
type CreateStockRequest struct {
	// AlcoholPercentage アルコール度数
	AlcoholPercentage *SakeAlcoholPercentage `json:"alcoholPercentage"`

	// Category 酒カテゴリ
	Category SakeCategory `json:"category"`

	// ImageKeys 表示順で指定する画像キー配列
	ImageKeys []SakeImageKey `json:"imageKeys"`

	// Memo メモ
	Memo *SakeMemo `json:"memo"`

	// Name 酒名
	Name SakeName `json:"name"`

	// Phonetic ふりがな
	Phonetic *SakePhonetic `json:"phonetic"`

	// Price 価格（円）
	Price *SakePrice `json:"price"`

	// Region 産地
	Region *SakeRegion `json:"region"`

	// TagNames タグ名前一覧
	TagNames []SakeTagName `json:"tagNames"`

	// VolumeMax 最大容量（ml）
	VolumeMax *SakeVolumeMax `json:"volumeMax"`

	// VolumeRemain 残量（%）
	VolumeRemain *SakeVolumeRemain `json:"volumeRemain"`
}

// DateTime 日時
type DateTime = time.Time

// ErrorResponse defines model for ErrorResponse.
type ErrorResponse struct {
	// Data エラー時は常に null
	Data *map[string]interface{} `json:"data,omitempty"`

	// Errors 発生したエラー情報の一覧
	Errors []APIError `json:"errors"`
}

// LikeCount いいね件数
type LikeCount = int

// ListSakesResponse お酒一覧取得APIのレスポンス
type ListSakesResponse struct {
	// Data 酒一覧情報
	Data []Sake `json:"data"`

	// Meta ページネーション情報
	Meta SakeListMeta `json:"meta"`
}

// ListStockResponse 在庫一覧取得APIのレスポンス
type ListStockResponse struct {
	// Data 在庫一覧情報
	Data []Stock `json:"data"`

	// Meta ページネーション情報
	Meta SakeListMeta `json:"meta"`
}

// Sake 酒情報
type Sake struct {
	// Category 酒カテゴリ
	Category SakeCategory `json:"category"`

	// FirstImageKey サムネイル画像のキー（存在する場合）
	FirstImageKey *string `json:"firstImageKey"`

	// LikeCount いいね件数
	LikeCount LikeCount `json:"likeCount"`

	// Name 酒名
	Name SakeName `json:"name"`

	// Phonetic ふりがな
	Phonetic *SakePhonetic `json:"phonetic"`

	// SakeId 酒ID
	SakeId SakeID `json:"sakeId"`
}

// SakeAlcoholPercentage アルコール度数
type SakeAlcoholPercentage = float32

// SakeCategory 酒カテゴリ
type SakeCategory string

// SakeDetail お酒の詳細情報
type SakeDetail struct {
	// AlcoholPercentage アルコール度数
	AlcoholPercentage *SakeAlcoholPercentage `json:"alcoholPercentage"`

	// Category 酒カテゴリ
	Category SakeCategory `json:"category"`

	// CreatedAt 日時
	CreatedAt DateTime `json:"createdAt"`

	// Images 画像情報（表示順に配列）
	Images []SakeImage `json:"images"`

	// LikeCount いいね件数
	LikeCount LikeCount `json:"likeCount"`

	// Memo メモ
	Memo *SakeMemo `json:"memo"`

	// Name 酒名
	Name SakeName `json:"name"`

	// Phonetic ふりがな
	Phonetic *SakePhonetic `json:"phonetic"`

	// Region 産地
	Region *SakeRegion `json:"region"`

	// SakeId 酒ID
	SakeId SakeID `json:"sakeId"`

	// Tags タグ情報（名前昇順に配列）
	Tags []SakeTag `json:"tags"`

	// UpdatedAt 日時
	UpdatedAt DateTime `json:"updatedAt"`
}

// SakeID 酒ID
type SakeID = openapi_types.UUID

// SakeImage defines model for SakeImage.
type SakeImage struct {
	// ImageId 酒イメージID
	ImageId openapi_types.UUID `json:"imageId"`

	// ImageKey 画像オブジェクトキー（S3などのストレージ上で一意な識別子）
	ImageKey SakeImageKey `json:"imageKey"`

	// SortOrder 画像表示順（0から始まる連番）
	SortOrder SakeSortOrder `json:"sortOrder"`
}

// SakeImageKey 画像オブジェクトキー（S3などのストレージ上で一意な識別子）
type SakeImageKey = string

// SakeListMeta defines model for SakeListMeta.
type SakeListMeta struct {
	// Limit 応答で適用された取得件数上限
	Limit int `json:"limit"`

	// Offset 応答で適用されたオフセット値
	Offset int `json:"offset"`

	// Total 条件に一致する総件数
	Total int `json:"total"`
}

// SakeMemo メモ
type SakeMemo = string

// SakeName 酒名
type SakeName = string

// SakePhonetic ふりがな
type SakePhonetic = string

// SakePrice 価格（円）
type SakePrice = int

// SakeRegion 産地
type SakeRegion = string

// SakeSortOrder 画像表示順（0から始まる連番）
type SakeSortOrder = int

// SakeTag defines model for SakeTag.
type SakeTag struct {
	// TagId タグID
	TagId SakeTagID `json:"tagId"`

	// TagName タグ名
	TagName SakeTagName `json:"tagName"`
}

// SakeTagID タグID
type SakeTagID = openapi_types.UUID

// SakeTagName タグ名
type SakeTagName = string

// SakeVolumeMax 最大容量（ml）
type SakeVolumeMax = int

// SakeVolumeRemain 残量（%）
type SakeVolumeRemain = int

// Stock 在庫情報
type Stock struct {
	// Category 酒カテゴリ
	Category SakeCategory `json:"category"`

	// FirstImageKey サムネイル画像のキー（存在する場合）
	FirstImageKey *string `json:"firstImageKey"`

	// LikeCount いいね件数
	LikeCount LikeCount `json:"likeCount"`

	// Name 酒名
	Name SakeName `json:"name"`

	// Phonetic ふりがな
	Phonetic *SakePhonetic `json:"phonetic"`

	// SakeId 酒ID
	SakeId SakeID `json:"sakeId"`
}

// StockDetail 在庫の詳細情報
type StockDetail struct {
	// AlcoholPercentage アルコール度数
	AlcoholPercentage *SakeAlcoholPercentage `json:"alcoholPercentage"`

	// Category 酒カテゴリ
	Category SakeCategory `json:"category"`

	// CreatedAt 日時
	CreatedAt DateTime `json:"createdAt"`

	// Images 画像情報（表示順に配列）
	Images []SakeImage `json:"images"`

	// LikeCount いいね件数
	LikeCount LikeCount `json:"likeCount"`

	// Memo メモ
	Memo *SakeMemo `json:"memo"`

	// Name 酒名
	Name SakeName `json:"name"`

	// Phonetic ふりがな
	Phonetic *SakePhonetic `json:"phonetic"`

	// Price 価格（円）
	Price *SakePrice `json:"price"`

	// Region 産地
	Region *SakeRegion `json:"region"`

	// SakeId 酒ID
	SakeId SakeID `json:"sakeId"`

	// Tags タグ情報（名前昇順に配列）
	Tags []SakeTag `json:"tags"`

	// UpdatedAt 日時
	UpdatedAt DateTime `json:"updatedAt"`

	// VolumeMax 最大容量（ml）
	VolumeMax *SakeVolumeMax `json:"volumeMax"`

	// VolumeRemain 残量（%）
	VolumeRemain *SakeVolumeRemain `json:"volumeRemain"`
}

// LikeToken defines model for LikeToken.
type LikeToken = string

// Limit defines model for Limit.
type Limit = int

// Offset defines model for Offset.
type Offset = int

// SakeId defines model for SakeId.
type SakeId = openapi_types.UUID

// BadRequest defines model for BadRequest.
type BadRequest = ErrorResponse

// Conflict defines model for Conflict.
type Conflict = ErrorResponse

// InternalServerError defines model for InternalServerError.
type InternalServerError = ErrorResponse

// NotFound defines model for NotFound.
type NotFound = ErrorResponse

// ServiceUnavailable defines model for ServiceUnavailable.
type ServiceUnavailable struct {
	Database struct {
		Postgres *ServiceUnavailableDatabasePostgres `json:"postgres,omitempty"`
		Valkey   *ServiceUnavailableDatabaseValkey   `json:"valkey,omitempty"`
	} `json:"database"`
	Status string `json:"status"`

	// Timestamp 応答生成時刻(UTC)
	Timestamp time.Time `json:"timestamp"`
}

// ServiceUnavailableDatabasePostgres defines model for ServiceUnavailable.Database.Postgres.
type ServiceUnavailableDatabasePostgres string

// ServiceUnavailableDatabaseValkey defines model for ServiceUnavailable.Database.Valkey.
type ServiceUnavailableDatabaseValkey string

// ListSakesParams defines parameters for ListSakes.
type ListSakesParams struct {
	// Offset スキップする件数
	Offset Offset `form:"offset" json:"offset"`

	// Limit 取得件数
	Limit Limit `form:"limit" json:"limit"`
}

// DeleteSakeLikeParams defines parameters for DeleteSakeLike.
type DeleteSakeLikeParams struct {
	// XLikeToken 匿名ユーザー識別用の一時トークン（UUID推奨）
	XLikeToken LikeToken `json:"X-Like-Token"`
}

// CreateSakeLikeParams defines parameters for CreateSakeLike.
type CreateSakeLikeParams struct {
	// XLikeToken 匿名ユーザー識別用の一時トークン（UUID推奨）
	XLikeToken LikeToken `json:"X-Like-Token"`
}

// ListStocksParams defines parameters for ListStocks.
type ListStocksParams struct {
	// Offset スキップする件数
	Offset Offset `form:"offset" json:"offset"`

	// Limit 取得件数
	Limit Limit `form:"limit" json:"limit"`
}

// CreateStockJSONRequestBody defines body for CreateStock for application/json ContentType.
type CreateStockJSONRequestBody = CreateStockRequest

// UpdateStockJSONRequestBody defines body for UpdateStock for application/json ContentType.
type UpdateStockJSONRequestBody = CreateStockRequest

// ServerInterface represents all server handlers.
type ServerInterface interface {
	// ヘルスチェック
	// (GET /health)
	HealthCheck(c *gin.Context)
	// 酒一覧取得
	// (GET /sakes)
	ListSakes(c *gin.Context, params ListSakesParams)
	// 酒詳細取得
	// (GET /sakes/{sakeId})
	GetSakeDetail(c *gin.Context, sakeId SakeId)
	// いいね解除
	// (DELETE /sakes/{sakeId}/likes)
	DeleteSakeLike(c *gin.Context, sakeId SakeId, params DeleteSakeLikeParams)
	// いいね登録
	// (POST /sakes/{sakeId}/likes)
	CreateSakeLike(c *gin.Context, sakeId SakeId, params CreateSakeLikeParams)
	// 在庫一覧取得
	// (GET /stocks)
	ListStocks(c *gin.Context, params ListStocksParams)
	// 在庫登録
	// (POST /stocks)
	CreateStock(c *gin.Context)
	// 在庫削除
	// (DELETE /stocks/{sakeId})
	DeleteStock(c *gin.Context, sakeId SakeId)
	// 在庫詳細取得
	// (GET /stocks/{sakeId})
	GetStockDetail(c *gin.Context, sakeId SakeId)
	// 在庫更新
	// (PUT /stocks/{sakeId})
	UpdateStock(c *gin.Context, sakeId SakeId)
}

// ServerInterfaceWrapper converts contexts to parameters.
type ServerInterfaceWrapper struct {
	Handler            ServerInterface
	HandlerMiddlewares []MiddlewareFunc
	ErrorHandler       func(*gin.Context, error, int)
}

type MiddlewareFunc func(c *gin.Context)

// HealthCheck operation middleware
func (siw *ServerInterfaceWrapper) HealthCheck(c *gin.Context) {

	for _, middleware := range siw.HandlerMiddlewares {
		middleware(c)
		if c.IsAborted() {
			return
		}
	}

	siw.Handler.HealthCheck(c)
}

// ListSakes operation middleware
func (siw *ServerInterfaceWrapper) ListSakes(c *gin.Context) {

	var err error

	// Parameter object where we will unmarshal all parameters from the context
	var params ListSakesParams

	// ------------- Required query parameter "offset" -------------

	if paramValue := c.Query("offset"); paramValue != "" {

	} else {
		siw.ErrorHandler(c, fmt.Errorf("Query argument offset is required, but not found"), http.StatusBadRequest)
		return
	}

	err = runtime.BindQueryParameterWithOptions("form", true, true, "offset", c.Request.URL.Query(), &params.Offset, runtime.BindQueryParameterOptions{Type: "integer", Format: ""})
	if err != nil {
		siw.ErrorHandler(c, fmt.Errorf("Invalid format for parameter offset: %w", err), http.StatusBadRequest)
		return
	}

	// ------------- Required query parameter "limit" -------------

	if paramValue := c.Query("limit"); paramValue != "" {

	} else {
		siw.ErrorHandler(c, fmt.Errorf("Query argument limit is required, but not found"), http.StatusBadRequest)
		return
	}

	err = runtime.BindQueryParameterWithOptions("form", true, true, "limit", c.Request.URL.Query(), &params.Limit, runtime.BindQueryParameterOptions{Type: "integer", Format: ""})
	if err != nil {
		siw.ErrorHandler(c, fmt.Errorf("Invalid format for parameter limit: %w", err), http.StatusBadRequest)
		return
	}

	for _, middleware := range siw.HandlerMiddlewares {
		middleware(c)
		if c.IsAborted() {
			return
		}
	}

	siw.Handler.ListSakes(c, params)
}

// GetSakeDetail operation middleware
func (siw *ServerInterfaceWrapper) GetSakeDetail(c *gin.Context) {

	var err error

	// ------------- Path parameter "sakeId" -------------
	var sakeId SakeId

	err = runtime.BindStyledParameterWithOptions("simple", "sakeId", c.Param("sakeId"), &sakeId, runtime.BindStyledParameterOptions{Explode: false, Required: true, Type: "string", Format: "uuid"})
	if err != nil {
		siw.ErrorHandler(c, fmt.Errorf("Invalid format for parameter sakeId: %w", err), http.StatusBadRequest)
		return
	}

	for _, middleware := range siw.HandlerMiddlewares {
		middleware(c)
		if c.IsAborted() {
			return
		}
	}

	siw.Handler.GetSakeDetail(c, sakeId)
}

// DeleteSakeLike operation middleware
func (siw *ServerInterfaceWrapper) DeleteSakeLike(c *gin.Context) {

	var err error

	// ------------- Path parameter "sakeId" -------------
	var sakeId SakeId

	err = runtime.BindStyledParameterWithOptions("simple", "sakeId", c.Param("sakeId"), &sakeId, runtime.BindStyledParameterOptions{Explode: false, Required: true, Type: "string", Format: "uuid"})
	if err != nil {
		siw.ErrorHandler(c, fmt.Errorf("Invalid format for parameter sakeId: %w", err), http.StatusBadRequest)
		return
	}

	// Parameter object where we will unmarshal all parameters from the context
	var params DeleteSakeLikeParams

	headers := c.Request.Header

	// ------------- Required header parameter "X-Like-Token" -------------
	if valueList, found := headers[http.CanonicalHeaderKey("X-Like-Token")]; found {
		var XLikeToken LikeToken
		n := len(valueList)
		if n != 1 {
			siw.ErrorHandler(c, fmt.Errorf("Expected one value for X-Like-Token, got %d", n), http.StatusBadRequest)
			return
		}

		err = runtime.BindStyledParameterWithOptions("simple", "X-Like-Token", valueList[0], &XLikeToken, runtime.BindStyledParameterOptions{ParamLocation: runtime.ParamLocationHeader, Explode: false, Required: true, Type: "string", Format: ""})
		if err != nil {
			siw.ErrorHandler(c, fmt.Errorf("Invalid format for parameter X-Like-Token: %w", err), http.StatusBadRequest)
			return
		}

		params.XLikeToken = XLikeToken

	} else {
		siw.ErrorHandler(c, fmt.Errorf("Header parameter X-Like-Token is required, but not found"), http.StatusBadRequest)
		return
	}

	for _, middleware := range siw.HandlerMiddlewares {
		middleware(c)
		if c.IsAborted() {
			return
		}
	}

	siw.Handler.DeleteSakeLike(c, sakeId, params)
}

// CreateSakeLike operation middleware
func (siw *ServerInterfaceWrapper) CreateSakeLike(c *gin.Context) {

	var err error

	// ------------- Path parameter "sakeId" -------------
	var sakeId SakeId

	err = runtime.BindStyledParameterWithOptions("simple", "sakeId", c.Param("sakeId"), &sakeId, runtime.BindStyledParameterOptions{Explode: false, Required: true, Type: "string", Format: "uuid"})
	if err != nil {
		siw.ErrorHandler(c, fmt.Errorf("Invalid format for parameter sakeId: %w", err), http.StatusBadRequest)
		return
	}

	// Parameter object where we will unmarshal all parameters from the context
	var params CreateSakeLikeParams

	headers := c.Request.Header

	// ------------- Required header parameter "X-Like-Token" -------------
	if valueList, found := headers[http.CanonicalHeaderKey("X-Like-Token")]; found {
		var XLikeToken LikeToken
		n := len(valueList)
		if n != 1 {
			siw.ErrorHandler(c, fmt.Errorf("Expected one value for X-Like-Token, got %d", n), http.StatusBadRequest)
			return
		}

		err = runtime.BindStyledParameterWithOptions("simple", "X-Like-Token", valueList[0], &XLikeToken, runtime.BindStyledParameterOptions{ParamLocation: runtime.ParamLocationHeader, Explode: false, Required: true, Type: "string", Format: ""})
		if err != nil {
			siw.ErrorHandler(c, fmt.Errorf("Invalid format for parameter X-Like-Token: %w", err), http.StatusBadRequest)
			return
		}

		params.XLikeToken = XLikeToken

	} else {
		siw.ErrorHandler(c, fmt.Errorf("Header parameter X-Like-Token is required, but not found"), http.StatusBadRequest)
		return
	}

	for _, middleware := range siw.HandlerMiddlewares {
		middleware(c)
		if c.IsAborted() {
			return
		}
	}

	siw.Handler.CreateSakeLike(c, sakeId, params)
}

// ListStocks operation middleware
func (siw *ServerInterfaceWrapper) ListStocks(c *gin.Context) {

	var err error

	// Parameter object where we will unmarshal all parameters from the context
	var params ListStocksParams

	// ------------- Required query parameter "offset" -------------

	if paramValue := c.Query("offset"); paramValue != "" {

	} else {
		siw.ErrorHandler(c, fmt.Errorf("Query argument offset is required, but not found"), http.StatusBadRequest)
		return
	}

	err = runtime.BindQueryParameterWithOptions("form", true, true, "offset", c.Request.URL.Query(), &params.Offset, runtime.BindQueryParameterOptions{Type: "integer", Format: ""})
	if err != nil {
		siw.ErrorHandler(c, fmt.Errorf("Invalid format for parameter offset: %w", err), http.StatusBadRequest)
		return
	}

	// ------------- Required query parameter "limit" -------------

	if paramValue := c.Query("limit"); paramValue != "" {

	} else {
		siw.ErrorHandler(c, fmt.Errorf("Query argument limit is required, but not found"), http.StatusBadRequest)
		return
	}

	err = runtime.BindQueryParameterWithOptions("form", true, true, "limit", c.Request.URL.Query(), &params.Limit, runtime.BindQueryParameterOptions{Type: "integer", Format: ""})
	if err != nil {
		siw.ErrorHandler(c, fmt.Errorf("Invalid format for parameter limit: %w", err), http.StatusBadRequest)
		return
	}

	for _, middleware := range siw.HandlerMiddlewares {
		middleware(c)
		if c.IsAborted() {
			return
		}
	}

	siw.Handler.ListStocks(c, params)
}

// CreateStock operation middleware
func (siw *ServerInterfaceWrapper) CreateStock(c *gin.Context) {

	for _, middleware := range siw.HandlerMiddlewares {
		middleware(c)
		if c.IsAborted() {
			return
		}
	}

	siw.Handler.CreateStock(c)
}

// DeleteStock operation middleware
func (siw *ServerInterfaceWrapper) DeleteStock(c *gin.Context) {

	var err error

	// ------------- Path parameter "sakeId" -------------
	var sakeId SakeId

	err = runtime.BindStyledParameterWithOptions("simple", "sakeId", c.Param("sakeId"), &sakeId, runtime.BindStyledParameterOptions{Explode: false, Required: true, Type: "string", Format: "uuid"})
	if err != nil {
		siw.ErrorHandler(c, fmt.Errorf("Invalid format for parameter sakeId: %w", err), http.StatusBadRequest)
		return
	}

	for _, middleware := range siw.HandlerMiddlewares {
		middleware(c)
		if c.IsAborted() {
			return
		}
	}

	siw.Handler.DeleteStock(c, sakeId)
}

// GetStockDetail operation middleware
func (siw *ServerInterfaceWrapper) GetStockDetail(c *gin.Context) {

	var err error

	// ------------- Path parameter "sakeId" -------------
	var sakeId SakeId

	err = runtime.BindStyledParameterWithOptions("simple", "sakeId", c.Param("sakeId"), &sakeId, runtime.BindStyledParameterOptions{Explode: false, Required: true, Type: "string", Format: "uuid"})
	if err != nil {
		siw.ErrorHandler(c, fmt.Errorf("Invalid format for parameter sakeId: %w", err), http.StatusBadRequest)
		return
	}

	for _, middleware := range siw.HandlerMiddlewares {
		middleware(c)
		if c.IsAborted() {
			return
		}
	}

	siw.Handler.GetStockDetail(c, sakeId)
}

// UpdateStock operation middleware
func (siw *ServerInterfaceWrapper) UpdateStock(c *gin.Context) {

	var err error

	// ------------- Path parameter "sakeId" -------------
	var sakeId SakeId

	err = runtime.BindStyledParameterWithOptions("simple", "sakeId", c.Param("sakeId"), &sakeId, runtime.BindStyledParameterOptions{Explode: false, Required: true, Type: "string", Format: "uuid"})
	if err != nil {
		siw.ErrorHandler(c, fmt.Errorf("Invalid format for parameter sakeId: %w", err), http.StatusBadRequest)
		return
	}

	for _, middleware := range siw.HandlerMiddlewares {
		middleware(c)
		if c.IsAborted() {
			return
		}
	}

	siw.Handler.UpdateStock(c, sakeId)
}

// GinServerOptions provides options for the Gin server.
type GinServerOptions struct {
	BaseURL      string
	Middlewares  []MiddlewareFunc
	ErrorHandler func(*gin.Context, error, int)
}

// RegisterHandlers creates http.Handler with routing matching OpenAPI spec.
func RegisterHandlers(router gin.IRouter, si ServerInterface) {
	RegisterHandlersWithOptions(router, si, GinServerOptions{})
}

// RegisterHandlersWithOptions creates http.Handler with additional options
func RegisterHandlersWithOptions(router gin.IRouter, si ServerInterface, options GinServerOptions) {
	errorHandler := options.ErrorHandler
	if errorHandler == nil {
		errorHandler = func(c *gin.Context, err error, statusCode int) {
			c.JSON(statusCode, gin.H{"msg": err.Error()})
		}
	}

	wrapper := ServerInterfaceWrapper{
		Handler:            si,
		HandlerMiddlewares: options.Middlewares,
		ErrorHandler:       errorHandler,
	}

	router.GET(options.BaseURL+"/health", wrapper.HealthCheck)
	router.GET(options.BaseURL+"/sakes", wrapper.ListSakes)
	router.GET(options.BaseURL+"/sakes/:sakeId", wrapper.GetSakeDetail)
	router.DELETE(options.BaseURL+"/sakes/:sakeId/likes", wrapper.DeleteSakeLike)
	router.POST(options.BaseURL+"/sakes/:sakeId/likes", wrapper.CreateSakeLike)
	router.GET(options.BaseURL+"/stocks", wrapper.ListStocks)
	router.POST(options.BaseURL+"/stocks", wrapper.CreateStock)
	router.DELETE(options.BaseURL+"/stocks/:sakeId", wrapper.DeleteStock)
	router.GET(options.BaseURL+"/stocks/:sakeId", wrapper.GetStockDetail)
	router.PUT(options.BaseURL+"/stocks/:sakeId", wrapper.UpdateStock)
}